<FEAT> (FE): add mbti page #1
